refactor: move resolve dynamic links to firebase file
